Add support for skipping TLS certificate verification
Some servers use expired TLS certificates. Connecting to these networks
is only possible by skipping TLS certificate verification. This feature
is also useful for connecting to servers with different hostnames
(although a specific option for setting the right hostname could be
added in the future).

This adds support for a flag in the service command `network create`
that lets users disable TLS certificate verification for a particular
network.