Fix saving `+draft/react` is msgstores
This commit fixes a bug where only TAGMSGs with a `+react` tag would be
stored, when a `+draft/react` tag is normative. I have kept messages
containing `+react` as well, as they will most likely be used in the
future.

I have also reversed the order of tag inspection, as I found the new
order to be more readable.