oauth2: Pass the usename to the introspection URL
When using oauth2 to do user authentication on an external service,
it helps for the service to have access to the username associated
to the token, in order to confirm whether the token is valid.

This can be used to have clients simply send their password, then
have soju send the token (password) + username to the auth service.