Add support for channel pinned/muted status through METADATA
First, this implements support for the metadata-2 specification,
work-in-progress by progval on ircv3-specifications [1].

The METADATA implementation is bouncer-only (namely, it does not
interact with any upstream METADATA).

The implementation is specific in that there are only 2 supported
keys: "pinned" and "muted". The user cannot set any other keys.
Those keys can only be "0" or "1", representing a boolean.

This enables downstreams to get/set two flags on each "message
target" (which can be either users or channels): pinned and muted.

- "pinned" is "1" when the channel should be pinned in the client,
  e.g. shown at the top of the channel list
- "muted" is "1" when the channel should be muted, e.g. have its
  notifications on non-highlights hidden

The actual specific client behavior for "pinned" and "muted" is not
defined here. This just enabels clients to get/set the flags.

[1]: https://github.com/ircv3/ircv3-specifications/pull/501